Removing data from sdata in each iteration.
`createSubset()` has possible logical bug. If you provide a signal in **signalList_sample.csv** which is not in mf4 file then you can see the effect. How to recreate the issue?

Provide following data in CSV where **XYZ** is not in mf4 file.

```
Name | Type | Signal  

Speed2D | Cont | Speed2D  

AngleRoll | Cont | AngleRoll

XYZ | Cont | XYZ

StationarySpeed | Cont | StationarySpeed
```


Now during execution when signal is **XYZ**, `sg_pos` will be empty and interpreter will skip the `for` loop but it will enter the next `if` condition as the `sdata` is not empty. Due to this, in the end of function `subsetList` will contain 4 signals with 2 signals having `name' as **XYZ**.

Resolution:
Before second for loop, remove content of `sdata`
```
sdata = None
```